+++ app/components/ProfileCard.tsx
@@ -2,27 +2,15 @@
 
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { TokenWithbalance, useTokens } from "../api/hooks/useTokens";
-import { useQRCode } from 'next-qrcode';
+import { useState } from "react";
+import { useTokens } from "../api/hooks/useTokens";
 
-
-import { TokenList } from "./TokenList";
 import { Swap } from "./Swap";
+import { Assets } from "./Assets";
 
 import { Button } from "@/components/ui/button";
-import {
-    Dialog,
-    DialogClose,
-    DialogContent,
-    DialogDescription,
-    DialogFooter,
-    DialogHeader,
-    DialogTitle,
-    DialogTrigger,
-} from "@/components/ui/dialog"
-import { Check, CopyIcon, Info } from "lucide-react";
-  
+import { Skeleton } from "@/components/ui/skeleton";
+import ProfileCardSkeleton from "./skeletons/ProfileCardSkeleton";
 
 
 type Tab = "tokens" | "send" | "add_funds" | "swap" | "withdraw"
@@ -43,10 +31,7 @@ export const ProfileCard = ({publicKey}: {
     const { tokenBalances, loading } = useTokens(publicKey);
 
     if (session.status === "loading") {
-        // TODO: replace with a skeleton
-        return <div>
-            Loading...
-        </div>
+        return <ProfileCardSkeleton />
     }
 
     if (!session.data?.user) {
@@ -88,44 +73,6 @@ function Warning() {
     </div>
 }
 
-function Assets({publicKey, tokenBalances, loading}: {
-    publicKey: string;
-    tokenBalances: {
-        totalBalance: number,
-        tokens: TokenWithbalance[]
-    } | null;
-    loading: boolean;
-}) {
-
-    if (loading) {
-        return "Loading..."
-    }
-
-    return <div className="text-slate-500">
-        <div className="mx-12 py-2">
-            Account assets
-        </div>
-        <div className="flex justify-between mx-12">
-            <div className="flex">
-                <div className="text-5xl font-bold text-black">
-                    ${tokenBalances?.totalBalance}
-                </div>
-                <div className="font-slate-500 font-bold text-3xl flex flex-col justify-end pb-0 pl-2">
-                    USD
-                </div>
-            </div>
-
-            <div>
-                <QrDialog publicKey={publicKey}/>
-            </div>
-        </div>
-
-        <div className="pt-4 bg-slate-50 p-12 mt-4">
-            <TokenList tokens={tokenBalances?.tokens || []} />
-        </div>
-    </div>
-}
-
 function Greeting({
     image, name
 }: {
@@ -137,93 +84,4 @@ function Greeting({
            Welcome back, {name}
         </div>
     </div>
-}
-
-function QrDialog({ publicKey } : { publicKey: string}) {
-    const { Canvas } = useQRCode();
-    const [copied, setCopied] = useState(false);
-
-    useEffect(() => {
-        if (copied) {
-            let timeout = setTimeout(() => {
-                setCopied(false)
-            }, 3000)
-            return () => {
-                clearTimeout(timeout);
-            }
-        }
-    }, [copied])
-
-    return (
-        <Dialog>
-            <DialogTrigger asChild>
-                <Button>
-                    Your wallet address
-                </Button>
-            </DialogTrigger>
-            <DialogContent>
-                <DialogHeader>
-                    <DialogTitle>
-                        <div className="text-center">
-                            Your Wallet Address
-                        </div>
-                    </DialogTitle>
-                    <DialogDescription>
-                        <div className="text-center">
-                            You can deposit crypto or NFTs into your account via this Solana wallet address:
-                        </div>
-                    </DialogDescription>
-                </DialogHeader>
-
-                <div className="flex flex-col justify-center items-center py-4">
-                    <Canvas
-                        text={publicKey}
-                        options={{
-                            width: 300,
-                        }}
-                    />
-                    <div 
-                        className="relative flex h-12 w-full cursor-pointer items-center justify-between rounded-full border border-grey-100 bg-white px-4 css-0"
-                    >
-                        <div className="inline-flex w-full text-center text-foreground css-0">
-                            <div className="w-full text-center text-foreground css-0">
-                                {publicKey.slice(0, 4)}
-                                <span className="text-gray-400">....</span>
-                                {publicKey.slice(-4)}
-                            </div>                                        
-                        </div>
-                        <button 
-                            className="flex items-center justify-center absolute right-[5px] top-[50%] translate-y-[-50%]"
-                            onClick={() => {
-                                navigator.clipboard.writeText(publicKey)
-                                setCopied(true)
-                            }}
-                        >
-                            
-                            {copied ? <Check/> : <CopyIcon/>}
-                        </button>
-                    </div>
-                    <div className="flex mt-3 items-center gap-x-3">
-                        <Info className="h-4 w-4"/>
-                        <span className="text-xs text-muted-foreground">Only send crypto to this address via the Solana network.</span>
-                    </div>
-                </div>
-
-                <DialogFooter>
-                    <div className="w-full flex justify-between">
-                        <Button
-                            variant="outline"
-                            size="lg"
-                            className="flex-1 mr-2"
-                        >
-                            <a href={`https://solscan.io/account/${publicKey}`} target="_blank">View on SolScan</a>
-                        </Button>
-                        <DialogClose asChild>
-                            <Button variant="outline" size="lg" className="flex-1 ml-2">Done</Button>
-                        </DialogClose>
-                    </div>
-                </DialogFooter>
-            </DialogContent>
-        </Dialog>
-    )
 }
